+++ testing/ocaml-caller/rust/src/lib.rs
@@ -3,7 +3,8 @@
 
 use ocaml_interop::{
     ocaml_alloc, ocaml_export, ocaml_unpack_polymorphic_variant, ocaml_unpack_variant, to_ocaml,
-    OCaml, OCamlBytes, OCamlFloat, OCamlInt, OCamlInt32, OCamlInt64, OCamlList, ToOCaml, ToRust,
+    OCaml, OCamlBytes, OCamlFloat, OCamlInt, OCamlInt32, OCamlInt64, OCamlList, OCamlRooted,
+    ToOCaml,
 };
 use std::{thread, time};
 
@@ -20,19 +21,19 @@ enum PolymorphicMovement {
 }
 
 ocaml_export! {
-    fn rust_twice(_cr, num: OCaml<OCamlInt>) -> OCaml<OCamlInt> {
-        let num: i64 = num.to_rust();
+    fn rust_twice(cr, num: &OCamlRooted<OCamlInt>) -> OCaml<OCamlInt> {
+        let num: i64 = num.to_rust(cr);
         unsafe { OCaml::of_i64_unchecked(num * 2) }
     }
 
-    fn rust_twice_boxed_i64(cr, num: OCaml<OCamlInt64>) -> OCaml<OCamlInt64> {
-        let num: i64 = num.to_rust();
+    fn rust_twice_boxed_i64(cr, num: &OCamlRooted<OCamlInt64>) -> OCaml<OCamlInt64> {
+        let num: i64 = num.to_rust(cr);
         let result = num * 2;
         ocaml_alloc!(result.to_ocaml(cr))
     }
 
-    fn rust_twice_boxed_i32(cr, num: OCaml<OCamlInt32>) -> OCaml<OCamlInt32> {
-        let num: i32 = num.to_rust();
+    fn rust_twice_boxed_i32(cr, num: &OCamlRooted<OCamlInt32>) -> OCaml<OCamlInt32> {
+        let num: i32 = num.to_rust(cr);
         let result = num * 2;
         ocaml_alloc!(result.to_ocaml(cr))
     }
@@ -41,8 +42,8 @@ ocaml_export! {
         num * num2
     }
 
-    fn rust_twice_boxed_float(cr, num: OCaml<OCamlFloat>) -> OCaml<OCamlFloat> {
-        let num: f64 = num.to_rust();
+    fn rust_twice_boxed_float(cr, num: &OCamlRooted<OCamlFloat>) -> OCaml<OCamlFloat> {
+        let num: f64 = num.to_rust(cr);
         let result = num * 2.0;
         ocaml_alloc!(result.to_ocaml(cr))
     }
@@ -51,10 +52,10 @@ ocaml_export! {
         num * 2.0
     }
 
-    fn rust_increment_bytes(cr, bytes: OCaml<OCamlBytes>, first_n: OCaml<OCamlInt>) -> OCaml<OCamlBytes> {
-        let first_n: i64 = first_n.to_rust();
+    fn rust_increment_bytes(cr, bytes: &OCamlRooted<OCamlBytes>, first_n: &OCamlRooted<OCamlInt>) -> OCaml<OCamlBytes> {
+        let first_n: i64 = first_n.to_rust(cr);
         let first_n = first_n as usize;
-        let mut vec: Vec<u8> = bytes.to_rust();
+        let mut vec: Vec<u8> = bytes.to_rust(cr);
 
         for i in 0..first_n {
             vec[i] += 1;
@@ -63,8 +64,8 @@ ocaml_export! {
         ocaml_alloc!(vec.to_ocaml(cr))
     }
 
-    fn rust_increment_ints_list(cr, ints: OCaml<OCamlList<OCamlInt>>) -> OCaml<OCamlList<OCamlInt>> {
-        let mut vec: Vec<i64> = ints.to_rust();
+    fn rust_increment_ints_list(cr, ints: &OCamlRooted<OCamlList<OCamlInt>>) -> OCaml<OCamlList<OCamlInt>> {
+        let mut vec: Vec<i64> = ints.to_rust(cr);
 
         for i in 0..vec.len() {
             vec[i] += 1;
@@ -73,44 +74,45 @@ ocaml_export! {
         ocaml_alloc!(vec.to_ocaml(cr))
     }
 
-    fn rust_make_tuple(cr, fst: OCaml<String>, snd: OCaml<OCamlInt>) -> OCaml<(String, OCamlInt)> {
-        let fst: String = fst.to_rust();
-        let snd: i64 = snd.to_rust();
+    fn rust_make_tuple(cr, fst: &OCamlRooted<String>, snd: &OCamlRooted<OCamlInt>) -> OCaml<(String, OCamlInt)> {
+        let fst: String = fst.to_rust(cr);
+        let snd: i64 = snd.to_rust(cr);
         let tuple = (fst, snd);
         ocaml_alloc!(tuple.to_ocaml(cr))
     }
 
-    fn rust_make_some(cr, value: OCaml<String>) -> OCaml<Option<String>> {
-        let value: String = value.to_rust();
+    fn rust_make_some(cr, value: &OCamlRooted<String>) -> OCaml<Option<String>> {
+        let value: String = value.to_rust(cr);
         let some_value = Some(value);
         ocaml_alloc!(some_value.to_ocaml(cr))
     }
 
-    fn rust_make_ok(cr, value: OCaml<OCamlInt>) -> OCaml<Result<OCamlInt, String>> {
-        let value: i64 = value.to_rust();
+    fn rust_make_ok(cr, value: &OCamlRooted<OCamlInt>) -> OCaml<Result<OCamlInt, String>> {
+        let value: i64 = value.to_rust(cr);
         let ok_value: Result<i64, String> = Ok(value);
         to_ocaml!(cr, ok_value)
     }
 
-    fn rust_make_error(cr, value: OCaml<String>) -> OCaml<Result<OCamlInt, String>> {
-        let value: String = value.to_rust();
+    fn rust_make_error(cr, value: &OCamlRooted<String>) -> OCaml<Result<OCamlInt, String>> {
+        let value: String = value.to_rust(cr);
         let error_value: Result<i64, String> = Err(value);
         to_ocaml!(cr, error_value)
     }
 
-    fn rust_sleep_releasing(cr, millis: OCaml<OCamlInt>) {
-        let millis: i64 = millis.to_rust();
+    fn rust_sleep_releasing(cr, millis: &OCamlRooted<OCamlInt>) {
+        let millis: i64 = millis.to_rust(cr);
         cr.releasing_runtime(|| thread::sleep(time::Duration::from_millis(millis as u64)));
         OCaml::unit()
     }
 
-    fn rust_sleep(cr, millis: OCaml<OCamlInt>) {
-        let millis: i64 = millis.to_rust();
+    fn rust_sleep(cr, millis: &OCamlRooted<OCamlInt>) {
+        let millis: i64 = millis.to_rust(cr);
         thread::sleep(time::Duration::from_millis(millis as u64));
         OCaml::unit()
     }
 
-    fn rust_string_of_movement(cr, movement: OCaml<PolymorphicMovement>) -> OCaml<String> {
+    fn rust_string_of_movement(cr, movement: &OCamlRooted<PolymorphicMovement>) -> OCaml<String> {
+        let movement = cr.get(movement);
         let pm = ocaml_unpack_variant! {
             movement => {
                 Step(count: OCamlInt) => { Movement::Step {count} },
@@ -127,7 +129,8 @@ ocaml_export! {
         to_ocaml!(cr, s)
     }
 
-    fn rust_string_of_polymorphic_movement(cr, polymorphic_movement: OCaml<PolymorphicMovement>) -> OCaml<String> {
+    fn rust_string_of_polymorphic_movement(cr, polymorphic_movement: &OCamlRooted<PolymorphicMovement>) -> OCaml<String> {
+        let polymorphic_movement = cr.get(polymorphic_movement);
         let pm = ocaml_unpack_polymorphic_variant! {
             polymorphic_movement => {
                 Step(count: OCamlInt) => { PolymorphicMovement::Step {count} },